Get rid of options.cpp.
