+++ console/optionparsing/parseoptions.hpp
@@ -1,5 +1,5 @@
 #include <string>
-#include "../../lib/options/options.hpp"
+#include "../../lib/options/user_options.hpp"
 #include "selection_types.hpp"
 
 enum class mode
