+++ lib/options/CMakeLists.txt
@@ -1,7 +1,5 @@
 target_sources_local(options
         PRIVATE
-        options.cpp
-        options.hpp
         global_options.hpp
         global_options.cpp
         user_options.hpp
