+++ lib/util/util.hpp
@@ -2,7 +2,15 @@
 #define _MSYNC_UTIL_HPP_
 #include <string>
 #include <string_view>
+#include <optional>
 
 std::string make_api_url(const std::string_view instance_url, const std::string_view api_route);
 
+struct parsed_account
+{
+    std::string username;
+    std::string instance;
+};
+
+std::optional<parsed_account> parse_account_name(const std::string& name);
 #endif
